// frontend/src/api/model/normativeLevel1NodeDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NormativeLevel2NodeDto } from './normativeLevel2NodeDto.ts';

export interface NormativeLevel1NodeDto {
  id?: string;
  name?: string;
  label?: string;
  order?: number;
  description?: string;
  status?: string;
  level2Nodes?: NormativeLevel2NodeDto[];
}