// frontend/src/api/model/normativeLevel2NodeDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NormativeLevel3NodeDto } from './normativeLevel3NodeDto.ts';

export interface NormativeLevel2NodeDto {
  id?: string;
  name?: string;
  label?: string;
  order?: number;
  description?: string;
  level3Nodes?: NormativeLevel3NodeDto[];
}